Feat: 이미지 리사이징 기능 추가
